GuestHost/SharedClipboard/x11: schedule synchronisation mechanism removal. bugref:9041: clipboard sharing on macOS host is brittle, Guest->Host direction
The GuestHost part of the X11 shared clipboard code contains a synchronisation
mechanism in the form of a flag which is set before an X11 clipboard request
is triggered and cleared when it completes.  This is because the Xt toolkit,
which we use for the shared clipboard, does not support multiple clipboard
operations at once.  However, we run all X11 clipboard code on a single thread,
and any function which sets the flag also waits for the operation to complete,
so the flag should not be necessary at all.  What is more, if there is any
valid use for it it is far from clear that it would actually work as intended.
This change schedules it for removal after 5.2 using conditional compilation.