Main/Machine: IMachine::mountMedium inadvertently works without a session. bugref:5490
IMachine::mountMedium() and IMachine::unmountMedium() need to detect and
reject attempts to use an immutable Machine object rather than a mutable
session object from ISession::machine().  This is done by including a
call to i_checkStateDependency(MutableOrRunningStateDep) as is done
elsewhere for IMachine methods that modify a VM's configuration or
settings.