Audio/WAS: Attempt to fix crashes when invalidating the cached audio interfaces of the device entry configurations. This is needed when the audio interface becomes invalid on a host device switch.
This patch modifies the original one so that we don't remove the current device entry from the cache, but just create a new cached entry with the new audio interface instead. The former approach apparently caused crashes when putting back a stale cache entry back to the cache when completing the device switch.

bugref:10503